Renamed classes, added javadoc and got rid of internationalization #554
