--- core/RhythmTestMocks/src/main/resources/org/jjazzlab/rhythmstubssimple/Bundle.properties
@@ -1,2 +1,0 @@
-DUMMY_RHYTHMS=Dummy rhythms
-DUMMY_RHYTHMS_DESC=Provides a dummy rhythm for each time signature
